+++ mex/editor/components.py
@@ -38,23 +38,29 @@ def render_link(value: EditorValue) -> rx.Component:
     )
 
 
+def render_span(text: str | None) -> rx.Component:
+    """Render a generic span with the given text."""
+    return rx.text(
+        text,
+        as_="span",
+    )
+
+
 def render_text(value: EditorValue) -> rx.Component:
     """Render an editor value as a text span."""
     return rx.skeleton(
-        rx.text(
-            cast("rx.vars.StringVar", value.text) | "Loading ...",
-            as_="span",
-        ),
+        render_span(cast("rx.vars.StringVar", value.text) | "Loading ..."),
         loading=~cast("rx.vars.StringVar", value.text),
     )
 
 
-def render_badge(value: EditorValue) -> rx.Component:
-    """Render a generic badge for an editor value."""
+def render_badge(text: str | None) -> rx.Component:
+    """Render a generic badge with the given text."""
     return rx.badge(
-        value.badge,
+        text,
         radius="large",
         variant="soft",
+        color_scheme="gray",
         style={"margin": "auto 0"},
     )
 
@@ -69,7 +75,7 @@ def render_value(value: EditorValue) -> rx.Component:
         ),
         rx.cond(
             value.badge,
-            render_badge(value),
+            render_badge(value.badge),
         ),
         spacing="1",
     )