+++ mex/editor/edit/state.py
@@ -127,10 +127,21 @@ class EditState(State):
     def _get_primary_sources_by_field_name(
         self, field_name: str
     ) -> list[EditorPrimarySource]:
+        """Get all primary sources for the given field name."""
         for field in self.fields:
             if field.name == field_name:
                 return field.primary_sources
         msg = f"field not found: {field_name}"
+        raise ValueError(msg)
+
+    def _get_editable_primary_source_by_field_name(
+        self, field_name: str
+    ) -> EditorPrimarySource:
+        """Get the (first) primary source that allows an editable rule."""
+        for primary_source in self._get_primary_sources_by_field_name(field_name):
+            if primary_source.input_config.allow_additive:
+                return primary_source
+        msg = f"editable field not found: {field_name}"
         raise ValueError(msg)
 
     @rx.event
@@ -161,35 +172,30 @@ class EditState(State):
     @rx.event
     def add_additive_value(self, field_name: str) -> None:
         """Add an additive rule to the given field."""
-        for primary_source in self._get_primary_sources_by_field_name(field_name):
-            if primary_source.input_config:
-                primary_source.editor_values.append(EditorValue())
+        primary_source = self._get_editable_primary_source_by_field_name(field_name)
+        primary_source.editor_values.append(EditorValue())
 
     @rx.event
     def remove_additive_value(self, field_name: str, index: int) -> None:
         """Remove an additive rule from the given field."""
-        for primary_source in self._get_primary_sources_by_field_name(field_name):
-            if primary_source.input_config:
-                primary_source.editor_values.pop(index)
+        primary_source = self._get_editable_primary_source_by_field_name(field_name)
+        primary_source.editor_values.pop(index)
 
     @rx.event
     def set_text_value(self, field_name: str, index: int, value: str) -> None:
         """Set the text attribute on an additive editor value."""
-        for primary_source in self._get_primary_sources_by_field_name(field_name):
-            if primary_source.input_config:
-                primary_source.editor_values[index].text = value
+        primary_source = self._get_editable_primary_source_by_field_name(field_name)
+        primary_source.editor_values[index].text = value
 
     @rx.event
     def set_badge_value(self, field_name: str, index: int, value: str) -> None:
         """Set the badge attribute on an additive editor value."""
-        for primary_source in self._get_primary_sources_by_field_name(field_name):
-            if primary_source.input_config:
-                primary_source.editor_values[index].badge = value
+        primary_source = self._get_editable_primary_source_by_field_name(field_name)
+        primary_source.editor_values[index].badge = value
 
     @rx.event
     def set_href_value(self, field_name: str, index: int, value: str) -> None:
         """Set an external href on an additive editor value."""
-        for primary_source in self._get_primary_sources_by_field_name(field_name):
-            if primary_source.input_config:
-                primary_source.editor_values[index].href = value
-                primary_source.editor_values[index].external = True
+        primary_source = self._get_editable_primary_source_by_field_name(field_name)
+        primary_source.editor_values[index].href = value
+        primary_source.editor_values[index].external = True