+++ mex/editor/transform.py
@@ -30,13 +30,13 @@ def transform_value(value: object, allow_link: bool = True) -> EditorValue:
     if isinstance(value, Text):
         return EditorValue(
             text=value.value,
-            badge=value.language,
+            badge=value.language.name if value.language else None,
         )
     if isinstance(value, Link):
         return EditorValue(
             text=value.title or value.url,
             href=value.url if allow_link else None,
-            badge=value.language,
+            badge=value.language.name if value.language else None,
             external=True,
         )
     if isinstance(value, Identifier):
@@ -46,8 +46,8 @@ def transform_value(value: object, allow_link: bool = True) -> EditorValue:
         )
     if isinstance(value, VocabularyEnum):
         return EditorValue(
-            text=value.name,
-            badge=type(value).__name__,
+            text=type(value).__name__,
+            badge=value.name,
         )
     if isinstance(value, TemporalEntity):
         return EditorValue(